+++ kaiba/functions.py
@@ -1,42 +1,37 @@
 import re
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, List, Optional, Union
 
 from returns.pipeline import flow
 from returns.pointfree import bind
-from returns.result import Failure, ResultE, Success, safe
+from returns.result import Failure, ResultE, safe
 
 from kaiba.casting import get_casting_function
-from kaiba.constants import (  # noqa: WPS235
-    CONDITION,
-    CONTAINS,
-    DEFAULT_GROUP,
-    FROM,
-    GROUP,
-    IN,
-    IS,
-    NOT,
-    ORIGINAL_FORMAT,
-    OTHERWISE,
-    SEARCH,
-    TARGET,
-    THEN,
-    TO,
+from kaiba.pydantic_schema import (
+    AnyType,
+    Casting,
+    ConditionEnum,
+    IfStatement,
+    Regexp,
+    Slicing,
 )
-from kaiba.valuetypes import MapValue, NewValue, ValueTypes
+from kaiba.valuetypes import ValueTypes
 
 
 @safe
-def apply_if_statements(if_value, if_objects) -> MapValue:
+def apply_if_statements(
+    if_value: Optional[AnyType],
+    if_objects: List[IfStatement],
+) -> Optional[AnyType]:
     """Apply if statements to a value.
 
     :param if_value: The value to use when evaluating if statements
-    :type if_value: MapValue
+    :type if_value: AnyType
 
     :param if_objects: :term:`if_objects` collection of if operations
     :type if_objects: List[Dict[str, Any]]
 
     :return: Success/Failure containers
-    :rtype: MapValue
+    :rtype: AnyType
 
     one if object looks like this
 
@@ -55,17 +50,21 @@ def apply_if_statements(if_value, if_objects) -> MapValue:
 
     Example
         >>> apply_if_statements(
-        ...     '1', [{'condition': 'is', 'target': '1', 'then': '2'}],
+        ...     '1', [
+        ...         IfStatement(
+        ...             **{'condition': 'is', 'target': '1', 'then': '2'}
+        ...         )
+        ...     ],
         ... ).unwrap() == '2'
         True
         >>> apply_if_statements(
         ...     'a',
-        ...     [{
+        ...     [IfStatement(**{
         ...         'condition': 'is',
         ...         'target': '1',
         ...         'then': '2',
         ...         'otherwise': '3'
-        ...     }],
+        ...     })],
         ... ).unwrap() == '3'
         True
 
@@ -74,7 +73,7 @@ def apply_if_statements(if_value, if_objects) -> MapValue:
 
         if_value = _apply_statement(
             if_value, if_object,
-        ).value_or(if_value)
+        )
 
     if if_value is None:
         raise ValueError('If statement failed or produced `None`')
@@ -82,51 +81,49 @@ def apply_if_statements(if_value, if_objects) -> MapValue:
     return if_value
 
 
-@safe
 def _apply_statement(
-    if_value: Optional[MapValue],
-    if_object: Dict[str, Any],
-) -> Optional[MapValue]:
+    if_value: Optional[AnyType],
+    if_object: IfStatement,
+) -> Optional[AnyType]:
     evaluation: bool = False
 
-    condition = if_object[CONDITION]
-    target = if_object[TARGET]
+    condition = if_object.condition
+    target = if_object.target
 
-    if condition == IS:
+    if condition == ConditionEnum.IS:
         evaluation = if_value == target
 
-    if condition == NOT:
+    if condition == ConditionEnum.NOT:
         evaluation = if_value != target
 
-    if condition == IN:
-        evaluation = if_value in target
+    if condition == ConditionEnum.IN:
+        evaluation = if_value in target  # type: ignore
 
-    if condition == CONTAINS:
+    if condition == ConditionEnum.CONTAINS:
         list_or_dict = isinstance(if_value, (dict, list))
         evaluation = list_or_dict and target in if_value  # type: ignore
-        evaluation = evaluation or not list_or_dict and target in str(if_value)
-
+        evaluation = evaluation or not list_or_dict and str(target) in str(if_value)  # noqa: E501 E262
     if evaluation:
-        return if_object[THEN]
+        return if_object.then
 
-    return if_object.get(OTHERWISE, if_value)
+    return if_object.otherwise or if_value
 
 
 @safe
 def apply_separator(
-    mapped_values: List[MapValue],
+    mapped_values: List[AnyType],
     separator: str,
-) -> MapValue:
+) -> AnyType:
     """Apply separator between the values of a List[Any].
 
     :param mapped_values: The list of values to join with the separator
-    :type mapped_values: List[MapValue]
+    :type mapped_values: List[AnyType]
 
     :param separator: :term:`separator` value to join mapped_values list with
     :type separator: str
 
     :return: Success/Failure containers
-    :rtype: MapValue
+    :rtype: AnyType
 
     Example
         >>> from returns.pipeline import is_successful
@@ -148,25 +145,25 @@ def apply_separator(
 
 
 def apply_slicing(
-    value_to_slice: Optional[NewValue],
-    slicing: Dict[str, Any],
-) -> Optional[NewValue]:
+    value_to_slice: Optional[Any],
+    slicing: Slicing,
+) -> Optional[AnyType]:
     """Slice value from index to index.
 
     :param value_to_slice: The value to slice
-    :type value_to_slice: MapValue
+    :type value_to_slice: Any
 
     :param slicing: :term:`slicing` object
     :type slicing: dict
 
 
     :return: Success/Failure containers
-    :rtype: MapValue
+    :rtype: Any
 
     Example
-        >>> apply_slicing('123', {'from': 1})
+        >>> apply_slicing('123', Slicing(**{'from': 1}))
         '23'
-        >>> apply_slicing('test', {'from': 1, 'to': 3})
+        >>> apply_slicing('test', Slicing(**{'from': 1, 'to': 3}))
         'es'
     """
     if value_to_slice is None:
@@ -178,66 +175,54 @@ def apply_slicing(
     if not isinstance(value_to_slice, list):
         value_to_slice = str(value_to_slice)
 
-    return value_to_slice[slicing[FROM]:slicing.get(TO)]
+    return value_to_slice[slicing.slice_from:slicing.slice_to]
 
 
 @safe
 def apply_regexp(  # noqa: WPS212, WPS234
-    value_to_match: Optional[MapValue],
-    regexp: Dict[str, Any],
-) -> Union[List[MapValue], MapValue, None]:
+    value_to_match: Optional[AnyType],
+    regexp: Regexp,
+) -> Union[List[AnyType], AnyType, None]:
     r"""Match value by a certain regexp pattern.
 
     :param value_to_match: The value to match
-    :type value_to_match: MapValue
+    :type value_to_match: AnyType
 
     :param regexp: :term: `matching` object which has parameters for
         regexp match
     :type regexp: dict
 
     :return: Success/Failure container
-    :rtype: MapValue
+    :rtype: AnyType
 
     Example
         >>> from kaiba.functions import apply_regexp
-        >>> apply_regexp('abcdef', {'search': '(?<=abc)def'}).unwrap()
+        >>> apply_regexp('abcdef', Regexp(**{'search': '(?<=abc)def'})).unwrap()
         'def'
         >>> apply_regexp(
         ...     'Isaac Newton, physicist',
-        ...     {'search': r'(\w+)', 'group': 1},
+        ...     Regexp(**{'search': r'(\w+)', 'group': 1}),
         ... ).unwrap()
         'Newton'
         >>> apply_regexp(
-        ...     'r7/p4pN1/1pn4k/8/2bP3R/2P3R1/6PP/6K1 b - -',
-        ...     {'search': r'(P[\d|\w])', 'group': [0, 2]},
+        ...     'Isaac Newton, physicist',
+        ...     Regexp(**{'search': r'(\w+)', 'group': [1, 2]}),
         ... ).unwrap()
-        ['P3', 'PP']
-        >>> apply_regexp(None, {'search': 'a+'}).unwrap()
-        >>> apply_regexp('lichess rocks!', {'search': None}).unwrap()
-        'lichess rocks!'
+        ['Newton', 'physicist']
+        >>> apply_regexp(None, Regexp(**{'search': 'a+'})).unwrap()
         >>> apply_regexp('Open-source matters', None).unwrap()
         'Open-source matters'
-        >>> apply_regexp(
-        ...     '1. e2 e4 d2 d4 2. Nc3 Nc6 3. Qe2 Qe7',
-        ...     {'search': r'(e\d)', 'group': []},
-        ... ).unwrap()
-        ['e2', 'e4', 'e2', 'e7']
-        >>> apply_regexp(
-        ...     '[Event \"Live Chess\"]\n[Site \"Chess.com\"]\n[Date ',
-        ...     {'search': r'Event \"[\d\w ]+\"'}
-        ... ).unwrap()
-        'Event "Live Chess"'
     """
     if value_to_match is None:
         return value_to_match
 
-    if not regexp or not regexp[SEARCH]:
+    if not regexp or not regexp.search:
         return value_to_match
 
-    pattern = regexp[SEARCH]
+    pattern = regexp.search
     groups = re.finditer(pattern, value_to_match)
     matches: list = [gr.group(0) for gr in groups]
-    num_group: Union[int, list] = regexp.get(GROUP, DEFAULT_GROUP)
+    num_group: Union[int, List[int]] = regexp.group
     if isinstance(num_group, list):
         if not num_group:
             return matches
@@ -246,38 +231,35 @@ def apply_regexp(  # noqa: WPS212, WPS234
 
 
 def apply_casting(
-    value_to_cast: Optional[MapValue],
-    casting: Dict[str, Any],
-) -> ResultE[MapValue]:
+    value_to_cast: Optional[AnyType],
+    casting: Casting,
+) -> ResultE[AnyType]:
     """Cast one type of code to another.
 
     :param casting: :term:`casting` object
     :type casting: dict
 
     :param value_to_cast: The value to cast to casting['to']
-    :type value_to_cast: MapValue
+    :type value_to_cast: AnyType
 
     :return: Success/Failure containers
-    :rtype: MapValue
+    :rtype: AnyType
 
     Example
-        >>> apply_casting('123', {'to': 'integer'}).unwrap()
+        >>> apply_casting('123', Casting(**{'to': 'integer'})).unwrap()
         123
-        >>> apply_casting('123.12', {'to': 'decimal'}).unwrap()
+        >>> apply_casting('123.12', Casting(**{'to': 'decimal'})).unwrap()
         Decimal('123.12')
     """
     if value_to_cast is None:
         return Failure(ValueError('value_to_cast is empty'))
 
-    if TO not in casting or casting[TO] is None:
-        return Success(value_to_cast)
-
     return flow(
-        casting[TO],
+        casting.to,
         get_casting_function,
         bind(  # type: ignore
             lambda function: function(  # type: ignore
-                value_to_cast, casting.get(ORIGINAL_FORMAT),
+                value_to_cast, casting.original_format,
             ),
         ),
     )
@@ -285,19 +267,19 @@ def apply_casting(
 
 @safe
 def apply_default(
-    mapped_value: Optional[Any] = None,
-    default: Optional[MapValue] = None,
-) -> MapValue:
+    mapped_value: Optional[AnyType] = None,
+    default: Optional[AnyType] = None,
+) -> AnyType:
     """Apply default value if exists and if mapped value is None.
 
     :param mapped_value: If this value is *None* the default value is returned
-    :type mapped_value: Optional[MapValue]
+    :type mapped_value: Optional[AnyType]
 
     :param default: :term:`default` value to return if mapped value is None
-    :type default: Optional[MapValue]
+    :type default: Optional[AnyType]
 
     :return: Success/Failure containers
-    :rtype: MapValue
+    :rtype: AnyType
 
     If default *is not* none and mapped_value *is* None then return default
     else if mapped_value is not in accepted ValueTypes then throw an error