+++ kaiba/process.py
@@ -1,42 +1,32 @@
-from typing import Callable, Union
+from typing import Union
 
-from attr import dataclass
+from pydantic import ValidationError
 from returns.functions import raise_exception
-from returns.pipeline import is_successful
-from returns.result import ResultE
-from typing_extensions import final
+from returns.result import Failure, ResultE
 
 from kaiba.mapper import map_data
-from kaiba.schema import SchemaValidator
-
-
-@final
-@dataclass(frozen=True, slots=True)
-class Process(object):
-    """Process Callable Object."""
-
-    validate: Callable[[dict], ResultE[dict]] = SchemaValidator()
-
-    def __call__(
-        self,
-        input_data: dict,
-        configuration: dict,
-    ) -> ResultE[Union[list, dict]]:
-        """Validate configuration then process data."""
-        cfg = self.validate(configuration)
-
-        if not is_successful(cfg):
-            return cfg
-
-        return map_data(input_data, cfg.unwrap())
+from kaiba.pydantic_schema import KaibaObject
 
 
 def process(
     input_data: dict,
     configuration: dict,
+) -> ResultE[Union[list, dict]]:
+    """Validate configuration then process data."""
+    try:
+        cfg = KaibaObject(**configuration)
+    except ValidationError as ve:
+        return Failure(ve)
+
+    return map_data(input_data, cfg)
+
+
+def process_raise(
+    input_data: dict,
+    configuration: dict,
 ) -> Union[list, dict]:
     """Call Process and unwrap value if no error, otherwise raise."""
-    return Process()(
+    return process(
         input_data,
         configuration,
     ).alt(