+++ tests/test_create_json_schema.py
@@ -1,0 +1,6 @@
+from kaiba.pydantic_schema import KaibaObject
+
+
+def test_create_jsonschema_from_model():
+    """Test that we can create jsonschema."""
+    assert KaibaObject.schema_json(indent=2)