+++ tests/test_nested_iterables.py
@@ -1,4 +1,5 @@
 from kaiba.collection_handlers import iterable_data_handler
+from kaiba.pydantic_schema import Iterable
 
 
 def test_iterable_data_handler():
@@ -21,22 +22,22 @@ def test_iterable_data_handler():
     }
 
     paths_to_iterables = [
-        {
+        Iterable(**{
             'alias': 'data',
             'path': ['data'],
-        },
-        {
+        }),
+        Iterable(**{
             'alias': 'nested',
             'path': ['data', 'nested'],
-        },
-        {
+        }),
+        Iterable(**{
             'alias': 'doesnotexist',
             'path': ['does', 'not', 'exist'],
-        },
-        {
+        }),
+        Iterable(**{
             'alias': 'another',
             'path': ['nested', 'another'],
-        },
+        }),
     ]
 
     iterables = iterable_data_handler(input_data, paths_to_iterables).unwrap()
@@ -47,4 +48,4 @@ def test_iterable_data_handler():
 def test_iterable_no_paths_returns_failure():
     """Test that when there are no paths we get a Failure."""
     iterables = iterable_data_handler({}, [])
-    assert 'No paths' in str(iterables.failure())
+    assert 'No iterables' in str(iterables.failure())